[expo-notifications] Rename Firebase service to NotificationsService
