+++ android/expoview/src/main/java/host/exp/exponent/fcm/ExpoFcmMessagingService.java
@@ -8,10 +8,10 @@ import org.json.JSONObject;
 import java.util.Map;
 
 import androidx.annotation.NonNull;
-import expo.modules.notifications.FirebaseListenerService;
 import expo.modules.notifications.notifications.model.NotificationContent;
 import expo.modules.notifications.notifications.model.NotificationRequest;
 import expo.modules.notifications.notifications.model.triggers.FirebaseNotificationTrigger;
+import expo.modules.notifications.service.NotificationsService;
 import host.exp.exponent.ABIVersion;
 import host.exp.exponent.Constants;
 import host.exp.exponent.analytics.EXL;
@@ -22,7 +22,7 @@ import host.exp.exponent.storage.ExperienceDBObject;
 import host.exp.exponent.storage.ExponentDB;
 
 
-public class ExpoFcmMessagingService extends FirebaseListenerService {
+public class ExpoFcmMessagingService extends NotificationsService {
 
   @Override
   public void onNewToken(@NonNull String token) {
